refactor: move auth module to modules
